make entity max health an int

+++ src/main/java/com/github/qeaml/mmic/mixin/DebugHudMixin.java
@@ -173,8 +173,8 @@ public class DebugHudMixin extends DrawableHelper {
 				Registry.ENTITY_TYPE.getId(ent.getType()),
 				ent.getPos().x, ent.getPos().y, ent.getPos().z));
 			if(ent instanceof LivingEntity lent)
-				cl.add(String.format("Health: %.2f/%.2f",
-					lent.getHealth(), lent.getAttributeValue(EntityAttributes.GENERIC_MAX_HEALTH)));
+				cl.add(String.format("Health: %.1f/%d",
+					lent.getHealth(), (int)lent.getAttributeValue(EntityAttributes.GENERIC_MAX_HEALTH)));
 			ent.getScoreboardTags().forEach(tag -> cl.add("#"+tag)); // FIXME: why doesn't this work?
 			int cbgw = 0;
 			for(String s: cl)
